refactor(124120): :hammer: filtros novo dashboard p&d
