+++ src/pages/DashboardGestaoProduto/DashboardGestaoProdutoPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Breadcrumb from "components/Shareable/Breadcrumb";
-import DashboardGestaoProduto from "components/screens/DashboardGestaoProduto";
+import { DashboardGestaoProduto } from "components/screens/DashboardGestaoProduto";
 import Page from "components/Shareable/Page/Page";
 
 const DashboardGestaoProdutoPage = (props) => {
